Cardano.ReCon.LTL.Check

Documentation

checkParamTerm :: Set PropVarIdentifier -> PropTerm -> [Error]

Validate a term by checking all of its free variables are bound.

checkParamConstraint :: Set PropVarIdentifier -> PropConstraint -> [Error]

Validate a constraint by ensuring its term is well-scoped.

checkFormula :: Set PropVarIdentifier -> Formula event ty -> [Error]

Check whether the formula is syntactically valid: | — all parameter variables shall be universally bound

data Error

Constructors

UnboundPropVarIdentifier PropVarIdentifier

prettyError :: Error -> Text