Cardano.ReCon.LTL.Lang.Formula

Documentation

type PropName = Text

A property name (e.g. "thread", "node", etc.).

type PropVarIdentifier = Text

Default name: x. | Identifier denoting an event property variable.

data PropValue

Default name: v. | An event property that can be either Int or Text.

Constructors

IntValue Int

TextValue Text

data PropTerm

Default name: t. | A term representing a constant property or a variable property.

Constructors

Const PropValue

Var PropVarIdentifier

data PropConstraint

Default name: c.

Constructors

PropConstraint PropName PropTerm

data Formula event ty

Default name: φ. | A type of Linear Temporal Logic formulas over a base type ty.

Constructors

Forall Word (Formula event ty)
☐ ᪲ₖ φ ≡ φ ∧ ◯ (◯ᵏ (☐ ᪲ₖ)) For every (k+1)-th unit of time from now, φ For example: ☐ ᪲₁ φ means for every other unit of time from now, φ ☐ ᪲₃ φ means for every 4-th unit of time from now, φ ☐ ᪲₀ φ means for every unit of time from now, φ

ForallN Word (Formula event ty)
☐ⁿ φ ☐⁰ φ ≡ ⊤ ☐¹⁺ᵏ φ ≡ φ ∧ ◯ (☐ᵏ φ) For each of the n units of time from now, φ

ExistsN Word (Formula event ty)
♢ⁿ φ ♢⁰ φ ≡ ⊥ ♢¹⁺ᵏ φ ≡ φ ∨ ◯ (♢ᵏ φ) For one of the n units of time from now, φ

Next (Formula event ty)
◯ φ For the next unit of time from now, φ.

NextN Word (Formula event ty)
◯ⁿ φ ◯⁰ φ ≡ φ ◯¹⁺ᵏ φ ≡ ◯ (◯ᵏ φ) For the n-th unit of time from now, φ

UntilN Word (Formula event ty) (Formula event ty)
φ |ⁿ ψ φ |⁰ ψ ≡ ⊤ φ |¹⁺ᵏ ψ ≡ ψ ∨ ¬ ψ ∧ φ ∧ (φ |ᵏ ψ) φ until ψ in the n units of time from now

Atom ty (Set PropConstraint)
ty c̄

Or (Formula event ty) (Formula event ty)
φ ∨ ψ

And (Formula event ty) (Formula event ty)
φ ∧ ψ

Not (Formula event ty)
¬ φ

Implies (Formula event ty) (Formula event ty)
φ ⇒ ψ

Top
T

Bottom
⊥

PropForall PropVarIdentifier (Formula event ty)
∀x. φ x implicitly ranges over the set of all integers and finite strings

PropForallN PropVarIdentifier (Set PropValue) (Formula event ty)
∀(x ∈ v̄). φ x ranges over values in `v̄`

PropEq (Relevance event ty) PropTerm PropValue
i = v

unfoldForall :: Word -> Formula event ty -> Formula event ty

unfoldForallN :: Word -> Formula event ty -> Formula event ty

unfoldExistsN :: Word -> Formula event ty -> Formula event ty

unfoldNextN :: Word -> Formula event ty -> Formula event ty

unfoldUntilN :: Word -> Formula event ty -> Formula event ty -> Formula event ty

relevance :: (Ord event, Ord ty) => Formula event ty -> Relevance event ty

Compute the total Relevance of the formula.

type Relevance event ty = Set (event, ty)

Set of relevant events.

and :: [Formula event ty] -> Formula event ty

interpTimeunit :: (Word -> Word) -> Formula event ty -> Formula event ty

It's useful to express temporal aspect of the formulas in a familiar time unit (e.g milliseconds). Yet, the LTL machinery works with nameless abstract time units. This function can be used to convert one into the other.

class Event a ty | a -> ty where

A constraint signifying that a is an Event over base ty: — Given an element of ty, ofTy shall name whether the event is of the given type. — Every event must have a distinct index (witnessed by index). — Every event of type ty (i.e. `ofTy event = True`) must have a key-value set of properties.

Methods

ofTy :: a -> ty -> Bool

Is the event of the given type?

props :: a -> ty -> Map PropVarIdentifier PropValue

Properties of the event pertinent to the given type. props e t assumes that ofTy e t = True

beg :: a -> Word64

Timestamp of the event in μs (Used for debug & monitoring only).