Cardano.ReCon.LTL.Rewrite

Documentation

rewriteHomogeneous :: Formula event ty -> Formula event ty

Rewrite the formula by applying the homogeneous fragment retraction & normalisation recursively.

rewriteFragment :: (Ord event, Ord ty) => Formula event ty -> Formula event ty

Rewrite the formula by applying the fragment retraction & normalisation recursively.

rewriteIdentity :: Eq ty => Formula event ty -> Formula event ty

Rewrites the formula by the following logical identities recursively: ☐ ᪲ₖ ⊤ = ⊤ ☐ᵏ ⊤ = ⊤ ♢ᵏ ⊥ = ⊥ φ |ᵏ ⊤ = ⊤ φ ∧ ⊤ = φ ⊤ ∧ φ = φ φ ∧ ⊥ = ⊥ ⊥ ∧ φ = ⊥ ⊤ ∨ φ = ⊤ φ ∨ ⊤ = ⊤ ⊥ ∨ φ = φ φ ∨ ⊥ = φ ⊤ ⇒ φ = φ ⊥ ⇒ φ = ⊤ φ ⇒ ⊤ = ⊤ φ ⇒ ⊥ = ¬ φ ¬ (¬ φ) = φ ¬ ⊥ = ⊤ ¬ ⊤ = ⊥ (v = v') = ⊤ where v = v' (v = v') = ⊥ where v ≠ v' ∀x. ⊤ = ⊤ ∀x. ⊥ = ⊥ ∀(x ∈ ∅). φ = ⊤ ∀(x ∈ {v} ⊔ v̄). ⊥ = ⊥ ∀(x ∈ {v} ⊔ v̄). ⊤ = ⊤ Additionally, unfolds base-cases of finite temporal operators.