Cardano.ReCon.LTL.Satisfy

Documentation

data SatisfactionResult event ty

The result of checking satisfaction of a formula against a timeline. | If unsatisfied, stores points in the timeline "relevant" to the formula.

Constructors

Satisfied

Unsatisfied (Relevance event ty)

satisfies :: (Event event ty, Ord event, Ord ty, Show ty, Foldable f) => Formula event ty -> f event -> SatisfactionResult event ty

Check if the formula is satisfied in the given event timeline.

satisfiesS :: (Event event ty, Ord event, Ord ty, Show ty) => Formula event ty -> Stream (Of event) IO () -> IORef (SatisfyMetrics event ty) -> IO (SatisfactionResult event ty)

Given a formula and a stream of events, forms an IO computation that returns a SatisfactionResult once the formula is equivalent to ⊤ or ⊥. This may happen either once the stream terminates or if the formula is falsified early by some prefix of the stream.

data SatisfyMetrics event ty

Constructors

SatisfyMetrics

Fields

eventsConsumed :: Word64

currentFormula :: Formula event ty

currentTimestamp :: Word64

μs