Cardano.ReCon.Trace.Feed

Documentation

data TemporalEvent

Temporal event represents multiple trace messages spanning some duration of time together with an index of the event.

Constructors

TemporalEvent

Fields

beg :: Word64

Microseconds since epoch when the event begins.

messages :: [TraceMessage]

type TemporalEventDurationMicrosec = Word

For performance considerations we group trace messages within the specified duration in one TemporalEvent.

read :: FilePath -> TemporalEventDurationMicrosec -> IO [TemporalEvent]

Read a text file where every line is a json object representation of a TraceMessage. Trace messages lying within the specified TemporalEventDurationMicrosec are grouped in TemporalEvent. The trace messages are sorted by timestamp before any action.

readS :: IngestorReader -> TemporalEventDurationMicrosec -> Stream (Of TemporalEvent) IO ()

Orphan instances

Eq TraceMessage

Instance details

Methods

(==) :: TraceMessage -> TraceMessage -> Bool

(/=) :: TraceMessage -> TraceMessage -> Bool

Ord TraceMessage

Instance details

Methods

compare :: TraceMessage -> TraceMessage -> Ordering

(<) :: TraceMessage -> TraceMessage -> Bool

(<=) :: TraceMessage -> TraceMessage -> Bool

(>) :: TraceMessage -> TraceMessage -> Bool

(>=) :: TraceMessage -> TraceMessage -> Bool

max :: TraceMessage -> TraceMessage -> TraceMessage

min :: TraceMessage -> TraceMessage -> TraceMessage